+++ Analytics.m
@@ -63,6 +63,16 @@ static NSString *GetSessionID() {
 #endif
 }
 
+static NSMutableDictionary *CreateContext(NSDictionary *parameters) {
+    NSMutableDictionary *context = [NSMutableDictionary dictionary];
+    [context setValue:@"analytics-ios-osx" forKey:@"library"];
+    // TODO add any device information here
+    if (parameters != nil) {
+        [context addEntriesFromDictionary:parameters];
+    }
+    return context;
+}
+
 
 
 
@@ -92,10 +102,20 @@ static Analytics *sharedAnalytics = nil;
 
 + (instancetype)sharedAnalyticsWithSecret:(NSString *)secret
 {
-    return [self sharedAnalyticsWithSecret:secret flushAt:20 flushAfter:30];
+    return [self sharedAnalyticsWithSecret:secret flushAt:20 flushAfter:30 delegate:nil];
+}
+
++ (instancetype)sharedAnalyticsWithSecret:(NSString *)secret delegate:(AnalyticsListenerDelegate *)delegate
+{
+    return [self sharedAnalyticsWithSecret:secret flushAt:20 flushAfter:30 delegate:delegate];
 }
 
 + (instancetype)sharedAnalyticsWithSecret:(NSString *)secret flushAt:(NSUInteger)flushAt flushAfter:(NSUInteger)flushAfter
+{
+    return [self sharedAnalyticsWithSecret:secret flushAt:flushAt flushAfter:flushAfter delegate:nil];
+}
+
++ (instancetype)sharedAnalyticsWithSecret:(NSString *)secret flushAt:(NSUInteger)flushAt flushAfter:(NSUInteger)flushAfter delegate:(AnalyticsListenerDelegate *)delegate
 {
     NSParameterAssert(secret.length > 0);
     NSParameterAssert(flushAt > 0);
@@ -103,7 +123,7 @@ static Analytics *sharedAnalytics = nil;
 
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
-        sharedAnalytics = [[self alloc] initWithSecret:secret flushAt:flushAt flushAfter:flushAfter];
+        sharedAnalytics = [[self alloc] initWithSecret:secret flushAt:flushAt flushAfter:flushAfter delegate:delegate];
     });
     return sharedAnalytics;
 }
@@ -114,13 +134,14 @@ static Analytics *sharedAnalytics = nil;
     return sharedAnalytics;
 }
 
-- (id)initWithSecret:(NSString *)secret flushAt:(NSUInteger)flushAt flushAfter:(NSUInteger)flushAfter
+- (id)initWithSecret:(NSString *)secret flushAt:(NSUInteger)flushAt flushAfter:(NSUInteger)flushAfter delegate:(AnalyticsListenerDelegate *)delegate
 {
     NSParameterAssert(secret.length);
     
     if (self = [self init]) {
         _flushAt = flushAt;
         _flushAfter = flushAfter;
+        _delegate = delegate;
         _secret = secret;
         _sessionId = GetSessionID();
         _queue = [NSMutableArray array];
@@ -141,10 +162,15 @@ static Analytics *sharedAnalytics = nil;
 
 - (void)identify:(NSString *)userId
 {
-    [self identify:userId traits:nil];
+    [self identify:userId traits:nil context:nil];
 }
 
 - (void)identify:(NSString *)userId traits:(NSDictionary *)traits
+{
+    [self identify:userId traits:traits context:nil];
+}
+
+- (void)identify:(NSString *)userId traits:(NSDictionary *)traits context:(NSDictionary *)context
 {
     dispatch_async(_serialQueue, ^{
         self.userId = userId;
@@ -153,16 +179,21 @@ static Analytics *sharedAnalytics = nil;
     NSMutableDictionary *dictionary = [NSMutableDictionary dictionary];
     [dictionary setValue:traits forKey:@"traits"];
 
-    [self enqueueAction:@"identify" dictionary:dictionary];
+    [self enqueueAction:@"identify" dictionary:dictionary context:context];
 }
 
 
 - (void)track:(NSString *)event
 {
-    [self track:event properties:nil];
+    [self track:event properties:nil context:nil];
 }
 
 - (void)track:(NSString *)event properties:(NSDictionary *)properties
+{
+    [self track:event properties:properties context:nil];
+}
+
+ - (void)track:(NSString *)event properties:(NSDictionary *)properties context:(NSDictionary *)context
 {
     NSAssert(event.length, @"%@ track requires an event name.", self);
 
@@ -170,14 +201,32 @@ static Analytics *sharedAnalytics = nil;
     [dictionary setValue:event forKey:@"event"];
     [dictionary setValue:properties forKey:@"properties"];
     
-    [self enqueueAction:@"track" dictionary:dictionary];
+    [self enqueueAction:@"track" dictionary:dictionary context:context];
+}
+
+
+- (void)alias:(NSString *)from to:(NSString *)to
+{
+    [self alias:from to:to context:nil];
+}
+
+- (void)alias:(NSString *)from to:(NSString *)to context:(NSDictionary *)context
+{
+    NSAssert(from.length, @"%@ alias requires a from id.", self);
+    NSAssert(to.length, @"%@ alias requires a to id.", self);
+
+    NSMutableDictionary *dictionary = [NSMutableDictionary dictionary];
+    [dictionary setValue:from forKey:@"from"];
+    [dictionary setValue:to forKey:@"to"];
+    
+    [self enqueueAction:@"alias" dictionary:dictionary context:context];
 }
 
 
 
 #pragma mark - Queueing
 
-- (void)enqueueAction:(NSString *)action dictionary:(NSMutableDictionary *)dictionary
+- (void)enqueueAction:(NSString *)action dictionary:(NSMutableDictionary *)dictionary context:(NSDictionary *)context
 {
     // attach these parts of the payload outside since they are all synchronous
     // and the timestamp will be more accurate.
@@ -185,6 +234,7 @@ static Analytics *sharedAnalytics = nil;
     [payload setValue:action forKey:@"action"];
     [payload setValue:ToISO8601([NSDate date]) forKey:@"timestamp"];
     [payload addEntriesFromDictionary:dictionary];
+    [payload setValue:CreateContext(context) forKey:@"context"];
 
     dispatch_async(_serialQueue, ^{
 
@@ -303,6 +353,10 @@ static Analytics *sharedAnalytics = nil;
         self.responseCode = 0;
         self.responseData = nil;
         self.connection = nil;
+
+        if (self.delegate) {
+            [self.delegate onAPISuccess];
+        }
     });
 }
 
@@ -315,6 +369,10 @@ static Analytics *sharedAnalytics = nil;
         self.responseCode = 0;
         self.responseData = nil;
         self.connection = nil;
+
+        if (self.delegate) {
+            [self.delegate onAPIFailure];
+        }
     });
 }