+++ AnalyticsTest/AnalyticsTest/main.m
@@ -1,0 +1,16 @@
+//
+//  main.m
+//  AnalyticsTest
+//
+//  Created by Peter Reinhardt on 4/10/13.
+//  Copyright (c) 2013 Segment.io. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+int main(int argc, char *argv[])
+{
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, @"GHUnitIOSAppDelegate");
+    }
+}