+++ AnalyticsTest/Pods/BuildHeaders/GHUnitIOS/GHTestViewModel.h
@@ -1,0 +1,1 @@
+../../GHUnitIOS/Classes/SharedUI/GHTestViewModel.h